refactor: migrate products feature to ADR-0001 structure (#191)
