+++ apps/carbotracker/src/features/products/+state/index.ts
@@ -1,0 +1,12 @@
+export * from './actions/api.actions';
+export * from './actions/component.actions';
+export * from './actions/dialog.actions';
+export * from './actions/routing.actions';
+export * from './actions/snackbar.actions';
+
+export * as apiEffects from './effects/api.effects';
+export * as dialogEffects from './effects/dialog.effects';
+export * as routingEffects from './effects/routing.effects';
+export * as snackbarEffects from './effects/snackbar.effects';
+
+export * from './products.store';
