+++ apps/carbotracker/src/features/products/product.model.ts
